backend/requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.2
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.20.0
Werkzeug==3.1.5
gunicorn
psycopg2-binary